# wrap emitter prototypes instead of instances where possible