### What does this PR do?
We can wrap methods on prototypes of EventEmitter subclasses in order to more quickly bind their instances.

Right now, only doing this with ServerResponse, as it's always in a hot path.

Later on, we may do this for all EventEmitters.
<!-- A brief description of the change being made with this pull request. -->

### Motivation
<!-- What inspired you to submit this pull request? -->
This boosts the performance of the `Scope#bind` operation, which happens, for
example, on every `req` and `res` object in HTTP server handlers.