@bengl 2020-10-21
Here are the benchmarks from CI. Note that on `master`, it's only effectively testing the same as `fake emitter` in this branch, but all cases would have similar performance due to how it was implemented.

The hottest path for us is `Scope#bind (subclass emitter)`.

```
--- [master]
Scope#bind (emitter) x 1,515,101 ops/sec ±5.51% (5 runs sampled)
Scope#bind (emitter.on/off) x 2,893,348 ops/sec ±5.96% (5 runs sampled)
--- [this branch/PR]
Scope#bind (fake emitter) x 1,637,104 ops/sec ±1.98% (5 runs sampled)
Scope#bind (fake emitter.on/off) x 3,320,290 ops/sec ±3.71% (5 runs sampled)
Scope#bind (emitter) x 1,848,739 ops/sec ±5.20% (5 runs sampled)
Scope#bind (emitter.on/off) x 2,085,796 ops/sec ±8.05% (5 runs sampled)
Scope#bind (subclass emitter) x 13,103,143 ops/sec ±2.77% (5 runs sampled)
Scope#bind (subclass emitter.on/off) x 1,807,546 ops/sec ±8.70% (5 runs sampled)
```