@bengl 2020-10-21
@rochdev I think that's exactly what `wrapEmitter` is doing, unless I'm misunderstanding.