@rochdev 2020-10-21
I'm basically referring to https://github.com/DataDog/dd-trace-js/pull/1126/files#diff-c93edee25fe3a71a04c788d4824ea8b4568e255d8de11ddc3547223f3c03859dR134

This should be done external to the scope manager. Basically, if the goal is to patch the prototype, then the prototype should be passed to `bindEmitter`.