@bengl 2020-10-22
@rochdev This is now re-worked so that `bindEmitter` is now a thing that's used internally on instances, and can be used externally (i.e. in plugins) to optimize hot-path EventEmitter subclasses.